# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TNC Nursing Classes API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Course {
  id: number;
  rowId: string;
  name: string;
  description: string;
  serialNo: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export type SessionContentType = typeof SessionContentType[keyof typeof SessionContentType];


export const SessionContentType = {
  youtube: 'youtube',
  firebase: 'firebase',
  pdf: 'pdf',
  none: 'none',
} as const;

export interface Session {
  id: number;
  rowId: string;
  title: string;
  description?: string;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  contentType: SessionContentType;
  type?: string;
  /** @nullable */
  courseId?: string | null;
  /** @nullable */
  subjectId?: string | null;
  isPaid: boolean;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  serialNo: string;
  createdAt?: string;
}

export interface Slider {
  id: number;
  rowId: string;
  imageUrl: string;
  name: string;
  description?: string;
}

export interface LoginInput {
  mobile: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  mobile: string;
  password: string;
  email?: string;
  college?: string;
  state?: string;
  city?: string;
  gender?: string;
  dob?: string;
}

export interface AuthResponse {
  userId: string;
  name: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  state?: string | null;
  token: string;
}

export interface Purchase {
  id: number;
  rowId: string;
  userId: string;
  courseId: string;
  courseName: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  paymentId?: string | null;
  createdAt?: string;
}

export interface PurchaseInput {
  userId: string;
  courseId: string;
  courseName: string;
  amount?: number;
  paymentId?: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResponse {
  token: string;
  message: string;
}

export interface AdminUserItem {
  id: number;
  rowId: string;
  name: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  state?: string | null;
  createdAt?: string;
}

export interface AdminStats {
  totalUsers: number;
  totalCourses: number;
  totalSessions: number;
  totalPurchases: number;
  recentUsers: AdminUserItem[];
}

export interface PromoStatus {
  enabled: boolean;
  /** @nullable */
  expiresAt?: string | null;
  message: string;
}

export interface PromoToggleInput {
  enabled: boolean;
  durationDays?: number;
  adminToken: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface StudyHeartbeatInput {
  telegramId: number;
  sessionId: string;
  seconds: number;
}

export interface StudyLeaderboardRow {
  telegramId: string;
  firstName: string;
  /** @nullable */
  username?: string | null;
  seconds: number;
}

export interface QuizMeta {
  examId: string;
  examNo: number;
  name: string;
  maxMarks: number;
  negativeMarks: number;
  durationMinutes: string;
  questionCount: number;
  /** @nullable */
  validUntil?: string | null;
  allowForPremium: boolean;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface QuizListResponse {
  quizzes: QuizMeta[];
  total: number;
  page: number;
  limit: number;
}

export interface QuizQuestion {
  rowId: string;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  questionNo?: number | null;
  questionId?: number;
}

export interface Quiz {
  examId: string;
  examNo?: number;
  name: string;
  maxMarks: number;
  negativeMarks: number;
  durationMinutes: string;
  questionCount?: number;
  /** @nullable */
  validUntil?: string | null;
  allowForPremium?: boolean;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  questions: QuizQuestion[];
}

export type ListSessionsParams = {
courseId?: string;
type?: string;
limit?: number;
};

export type ListQuizzesParams = {
page?: number;
limit?: number;
};