# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TNC Nursing Classes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/study/heartbeat:
    post:
      operationId: recordStudyHeartbeat
      tags: [study]
      summary: Record study time for a Telegram user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyHeartbeatInput"
      responses:
        "200":
          description: Study time recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid study heartbeat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    StudyHeartbeatInput:
      type: object
      properties:
        telegramId:
          type: integer
        sessionId:
          type: string
        seconds:
          type: number
      required:
        - telegramId
        - sessionId
        - seconds
...